+++ src/App.js
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import './App.css';
-import Header from './components/Header/Header';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import Header from './components/Header';
 
 const queryClient = new QueryClient();
 
@@ -9,7 +9,7 @@ function App() {
   return (
     <>
       <Header />
-      <main className="main">
+      <main className='main'>
         <QueryClientProvider client={queryClient}>
           <Outlet />
         </QueryClientProvider>
