+++ src/components/Header.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
+import styles from './css/Header.module.css';
 import { Link } from 'react-router-dom';
-import { useAuthContext } from '../../context/AuthContext';
-import styles from './Header.module.css';
-import Avatar from '../common/Avatar';
+import { useAuthContext } from '../context/AuthContext';
+import Avatar from './common/Avatar';
 
 export default function Header() {
   const { user, logout } = useAuthContext();
